Consolidate DecodeContent to vt/topo package - can't put in vt/topo/topoproto because of import cycle
