+++ go/vt/vtctl/grpcvtctldserver/topo.go
@@ -19,12 +19,7 @@ package grpcvtctldserver
 import (
 	"context"
 	"fmt"
-	"path"
 	"time"
-
-	"google.golang.org/protobuf/encoding/protojson"
-	"google.golang.org/protobuf/encoding/prototext"
-	"google.golang.org/protobuf/proto"
 
 	"vitess.io/vitess/go/trace"
 	"vitess.io/vitess/go/vt/log"
@@ -34,64 +29,8 @@ import (
 	"vitess.io/vitess/go/vt/vterrors"
 
 	topodatapb "vitess.io/vitess/go/vt/proto/topodata"
-	vschemapb "vitess.io/vitess/go/vt/proto/vschema"
 	"vitess.io/vitess/go/vt/proto/vtrpc"
 )
-
-// decodeContent uses the filename to imply a type, and proto-decodes
-// the right object, then echoes it as a string.
-//
-// This is copied from vtctl/topo. We cannot import vtctl/topo because of an import loop.
-func decodeContent(filename string, data []byte, json bool) (string, error) {
-	name := path.Base(filename)
-	dir := path.Dir(filename)
-	var p proto.Message
-	switch name {
-	case topo.CellInfoFile:
-		p = new(topodatapb.CellInfo)
-	case topo.KeyspaceFile:
-		p = new(topodatapb.Keyspace)
-	case topo.ShardFile:
-		p = new(topodatapb.Shard)
-	case topo.VSchemaFile:
-		p = new(vschemapb.Keyspace)
-	case topo.ShardReplicationFile:
-		p = new(topodatapb.ShardReplication)
-	case topo.TabletFile:
-		p = new(topodatapb.Tablet)
-	case topo.SrvVSchemaFile:
-		p = new(vschemapb.SrvVSchema)
-	case topo.SrvKeyspaceFile:
-		p = new(topodatapb.SrvKeyspace)
-	case topo.RoutingRulesFile:
-		p = new(vschemapb.RoutingRules)
-	default:
-		switch dir {
-		case "/" + topo.GetExternalVitessClusterDir():
-			p = new(topodatapb.ExternalVitessCluster)
-		default:
-		}
-		if p == nil {
-			if json {
-				return "", fmt.Errorf("unknown topo protobuf type for %v", name)
-			}
-			return string(data), nil
-		}
-	}
-
-	if err := proto.Unmarshal(data, p); err != nil {
-		return string(data), err
-	}
-
-	var marshalled []byte
-	var err error
-	if json {
-		marshalled, err = protojson.Marshal(p)
-	} else {
-		marshalled, err = prototext.Marshal(p)
-	}
-	return string(marshalled), err
-}
 
 func deleteShard(ctx context.Context, ts *topo.Server, keyspace string, shard string, recursive bool, evenIfServing bool, force bool) (err error) {
 	span, ctx := trace.NewSpan(ctx, "VtctldServer.deleteShard")
