+++ go/vt/vtctld/explorer.go
@@ -28,7 +28,6 @@ import (
 
 	"vitess.io/vitess/go/vt/topo"
 	"vitess.io/vitess/go/vt/topo/topoproto"
-	"vitess.io/vitess/go/vt/vtctl"
 )
 
 // backendExplorer is a class that uses the Backend interface of a
@@ -89,7 +88,7 @@ func (ex *backendExplorer) HandlePath(nodePath string, r *http.Request) *Result 
 	case nil:
 		if len(data) > 0 {
 			// It has contents, we just use it if possible.
-			decoded, err := vtctl.DecodeContent(relativePath, data, false)
+			decoded, err := topo.DecodeContent(relativePath, data, false)
 			if err != nil {
 				result.Error = err.Error()
 			} else {
